requirements.txt
fastapi
uvicorn[standard]
gitpython
github3.py
httpx
python-dotenv
requests 